requirements.txt
Flask==2.3.3
pycups==2.0.1
Werkzeug==2.3.7